Add continuous delivery workflow
Create a GitHub Actions workflow triggered on the `main` branch that
will automatically publish to npm. The idea is to just always try to
publish, which will only succeed if the version number in `package.json`
is updated. Otherwise the workflow _should_ (but I'm not 100% sure it
will) still succeed due to the `continue-on-error: true` directive.